Move delete_comments_and_strings and make_guide_lines to Importer class.
